requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
